\name{nucmapData-package}
\alias{nucmapData-package}
\alias{nucmapData}
\docType{package}
\title{
An example dataset for the NuCMap package
}
\description{
This experiment data package is for the use with the NuCMap package. It contains all the illustration data files. 
}
\author{
Liqun Xi, Qingyang Zhang, Kristin Brogaard, Bruce Lindsay, Jonathan Widom, Ji-Ping Wang
}
\examples{
library(nucmapData)
genfile1=system.file("extdata","chrI.fa",package="nucmapData")
}
